+++ android/src/main/kotlin/com/eopeter/fluttermapboxnavigation/models/MapBoxRouteStep.kt
@@ -1,13 +1,13 @@
 package com.eopeter.fluttermapboxnavigation.models
 
-import com.google.gson.JsonObject
 import com.mapbox.api.directions.v5.models.LegStep
+import com.google.gson.*
 
-class MapBoxRouteStep(private val step: LegStep) {
+class MapBoxRouteStep(val step: LegStep) {
 
-    private val instructions: String = "" //step.bannerInstructions()[0]?.primary().toString()
-    private val distance: Double = step.distance()
-    private val expectedTravelTime: Double = step.duration()
+    val instructions: String = "" //step.bannerInstructions()[0]?.primary().toString()
+    val distance: Double = step.distance()
+    val expectedTravelTime: Double = step.duration()
 
     fun toJsonObject(): JsonObject {
         val json = JsonObject()
